+++ lib/patches/active_record/serialization.rb
@@ -1,22 +1,5 @@
-module Globalize
-  module AttributeMethods
-    module Serialization
-      def serialize(attr_name, class_name_or_coder = Object)
-        super(attr_name, class_name_or_coder)
-
-        coder = if class_name_or_coder == ::JSON
-                  ::ActiveRecord::Coders::JSON
-                elsif [:load, :dump].all? { |x| class_name_or_coder.respond_to?(x) }
-                  class_name_or_coder
-                else
-                  ::ActiveRecord::Coders::YAMLColumn.new(class_name_or_coder)
-                end
-
-        self.globalize_serialized_attributes = globalize_serialized_attributes.dup
-        self.globalize_serialized_attributes[attr_name] = coder
-      end
-    end
-  end
+if ::ActiveRecord::VERSION::STRING < "5.1.0"
+  require_relative 'rails4/serialization'
+else
+  require_relative 'rails5_1/serialization'
 end
-
-ActiveRecord::AttributeMethods::Serialization::ClassMethods.send(:prepend, Globalize::AttributeMethods::Serialization)
