+++ test/data/models/serialized_attr.rb
@@ -3,7 +3,14 @@ class SerializedAttr < ActiveRecord::Base
   translates :meta
 end
 
-class JSONSerializedAttr < SerializedAttr
+class ArraySerializedAttr < ActiveRecord::Base
+  self.table_name = 'serialized_attrs'
+  serialize :meta, Array
+  translates :meta
+end
+
+class JSONSerializedAttr < ActiveRecord::Base
+  self.table_name = 'serialized_attrs'
   serialize :meta, JSON
   translates :meta
 end
